Move files and add UTF8 checks.
